# core-common-user-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-user-mgmt
  description: User Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/users
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/users/{userId}:
    get:
      tags:
        - User Accounts
      summary: Get user account by ID
      description: Returns a user account by its ID
      operationId: getUserAccountById
      parameters:
        - name: userId
          in: path
          description: ID of the user account to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved user account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
        '404':
          description: User account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
    put:
      tags:
        - User Accounts
      summary: Update an existing user account
      description: Updates an existing user account and returns the updated user account
      operationId: updateUserAccount
      parameters:
        - name: userId
          in: path
          description: ID of the user account to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAccountDTO'
        required: true
      responses:
        '200':
          description: User account successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
        '404':
          description: User account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
    delete:
      tags:
        - User Accounts
      summary: Delete a user account
      description: Deletes a user account by its ID
      operationId: deleteUserAccount
      parameters:
        - name: userId
          in: path
          description: ID of the user account to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User account successfully deleted
        '404':
          description: User account not found
...
  /api/v1/external-identities/{externalIdentityId}:
    get:
      tags:
        - User External Identities
      summary: Get user external identity by ID
      description: Returns a user external identity by its ID
      operationId: getUserExternalIdentityById
      parameters:
        - name: externalIdentityId
          in: path
          description: ID of the user external identity to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved user external identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserExternalIdentityDTO'
        '404':
          description: User external identity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserExternalIdentityDTO'
    put:
      tags:
        - User External Identities
      summary: Update an existing user external identity
      description: Updates an existing user external identity and returns the updated user external identity
      operationId: updateUserExternalIdentity
      parameters:
        - name: externalIdentityId
          in: path
          description: ID of the user external identity to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserExternalIdentityDTO'
        required: true
      responses:
        '200':
          description: User external identity successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserExternalIdentityDTO'
        '404':
          description: User external identity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserExternalIdentityDTO'
    delete:
      tags:
        - User External Identities
      summary: Delete a user external identity
      description: Deletes a user external identity by its ID
      operationId: deleteUserExternalIdentity
      parameters:
        - name: externalIdentityId
          in: path
          description: ID of the user external identity to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User external identity successfully deleted
        '404':
          description: User external identity not found
...
  /api/v1/users/{userId}/external-identities:
    get:
      tags:
        - User External Identities
      summary: Get external identities for a user
      description: Returns all external identities linked to a specific user
      operationId: getExternalIdentitiesByUserId
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved user external identities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseUserExternalIdentityDTO'
    post:
      tags:
        - User External Identities
      summary: Link external identity to user
      description: Links an external identity to a user and returns the created user external identity
      operationId: linkExternalIdentityToUser
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserExternalIdentityDTO'
        required: true
      responses:
        '201':
          description: External identity successfully linked to user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserExternalIdentityDTO'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserExternalIdentityDTO'
...
  /api/v1/users:
    post:
      tags:
        - User Accounts
      summary: Create a new user account
      description: Creates a new user account and returns the created user account
      operationId: createUserAccount
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAccountDTO'
        required: true
      responses:
        '201':
          description: User account successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
...
  /api/v1/external-identities/filter:
    post:
      tags:
        - User External Identities
      summary: Get all user external identities with filtering
      description: Returns a paginated list of user external identities based on filter criteria
      operationId: filterUserExternalIdentities
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestUserExternalIdentityDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved user external identities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
  /api/v1/external-identities:
    post:
      tags:
        - User External Identities
      summary: Create a new user external identity
      description: Creates a new user external identity and returns the created user external identity
      operationId: createUserExternalIdentity
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserExternalIdentityDTO'
        required: true
      responses:
        '201':
          description: User external identity successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserExternalIdentityDTO'
...
  /api/v1/audit-logs/users/{userId}:
    get:
      tags:
        - Audit Logs
      summary: Get audit logs for a user
      description: Returns all audit logs for a specific user
      operationId: getAuditLogsByUserId
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
...
  /api/v1/users/{userId}/external-identities/{externalIdentityId}:
    delete:
      tags:
        - User External Identities
      summary: Unlink external identity from user
      description: Unlinks an external identity from a user
      operationId: unlinkExternalIdentityFromUser
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: string
            format: uuid
        - name: externalIdentityId
          in: path
          description: ID of the external identity
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: External identity successfully unlinked from user
        '404':
          description: User external identity not found
...
components:
  schemas:
    UserAccountDTO:
      required:
        - email
        - fullName
        - isActive
        - userType
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        fullName:
          maxLength: 255
          minLength: 1
          type: string
        nickname:
          maxLength: 100
          minLength: 0
          type: string
        email:
          maxLength: 255
          minLength: 0
          type: string
        userType:
          type: string
          enum:
            - EMPLOYEE
            - DISTRIBUTOR
        branchId:
          type: string
          format: uuid
        distributorId:
          type: string
          format: uuid
        departmentId:
          type: string
          format: uuid
        positionId:
          type: string
          format: uuid
        jobTitle:
          maxLength: 255
          minLength: 0
          type: string
        avatarUrl:
          maxLength: 500
          minLength: 0
          pattern: ^(https?://)?[\w\-]+(\.[\w\-]+)+([\w\-\.,@?^=%&:/~\+#]*[\w\-\@?^=%&/~\+#])?$
          type: string
        themePreference:
          type: string
          enum:
            - LIGHT
            - DARK
            - SYSTEM
        languagePreference:
          maxLength: 10
          minLength: 0
          pattern: ^[a-z]{2}(-[A-Z]{2})?$
          type: string
        locale:
          maxLength: 10
          minLength: 0
          pattern: ^[a-z]{2}_[A-Z]{2}$
          type: string
        timezone:
          maxLength: 50
          minLength: 0
          type: string
        contactPhone:
          maxLength: 20
          minLength: 0
          pattern: ^\+?[1-9]\d{1,14}$
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    UserExternalIdentityDTO:
      required:
        - isPrimary
        - provider
        - subjectId
        - userAccountId
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        userAccountId:
          type: string
          format: uuid
        provider:
          maxLength: 100
          minLength: 1
          type: string
        subjectId:
          maxLength: 255
          minLength: 1
          type: string
        email:
          maxLength: 255
          minLength: 0
          type: string
        isPrimary:
          type: boolean
        linkedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestUserExternalIdentityDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/UserExternalIdentityDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
...
    PaginationResponseUserExternalIdentityDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/UserExternalIdentityDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...